feat: #49 시설 카테고리 관리 API 개발

+++ apps/cms-be/src/users/users.controller.ts
@@ -6,6 +6,8 @@ import {
   Patch,
   Param,
   Delete,
+  ConflictException,
+  NotFoundException,
 } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { Prisma } from '@prisma/client';
@@ -18,29 +20,43 @@ export class UsersController {
 
   @Post()
   async create(@Body() user: Prisma.UserCreateInput) {
+    console.log(user);
+    const sameUser = await this.usersService.findOne(user.userId);
+    console.log(sameUser);
+    if (sameUser) {
+      throw new ConflictException('Data already exist.');
+    }
     return this.usersService.create(user);
   }
 
   @Get()
-  findAll() {
+  async findAll() {
     return this.usersService.findAll();
   }
 
   @Get(':userId')
-  findOne(@Param('userId') userId: string) {
+  async findOne(@Param('userId') userId: string) {
     return this.usersService.findOne(userId);
   }
 
   @Patch(':userId')
-  update(
+  async update(
     @Param('userId') userId: string,
     @Body() data: Prisma.UserUpdateInput,
   ) {
+    const sameUser = await this.usersService.findOne(userId);
+    if (!sameUser) {
+      throw new NotFoundException('Data not found.');
+    }
     return this.usersService.update({ userId, data });
   }
 
   @Delete(':userId')
-  remove(@Param('userId') userId: string) {
+  async remove(@Param('userId') userId: string) {
+    const sameUser = await this.usersService.findOne(userId);
+    if (!sameUser) {
+      throw new NotFoundException('Data not found.');
+    }
     return this.usersService.remove(userId);
   }
 }
